[IN-323] chore: add helper.ts instead of js
