+++ src/shared/index.ts
@@ -2,3 +2,4 @@ export * from './utils'
 export * from './components'
 export * from './apolloClient'
 export * from './constants'
+export * from './types'
